+++ src/NeuralNet/Optimizers/AdaGrad.php
@@ -26,7 +26,7 @@ use const Rubix\ML\EPSILON;
 class AdaGrad implements Optimizer, Adaptive
 {
     /**
-     * The learning rate. i.e. the master step size.
+     * The learning rate that controls the global step size.
      *
      * @var float
      */