+++ src/NeuralNet/Optimizers/StepDecay.php
@@ -20,7 +20,7 @@ use InvalidArgumentException;
 class StepDecay implements Optimizer
 {
     /**
-     * The learning rate. i.e. the master step size.
+     * The learning rate that controls the global step size.
      *
      * @var float
      */