+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 - Unreleased
     - Added describe method to dataset object
+    - Added deduplicate method to dataset object
+    - Added unzip static factory for Labeled datasets from data table
     - Changed the order of t-SNE hyperparameters
 
 - 0.0.15-beta